+++ global.d.ts
@@ -1,0 +1,9 @@
+export { };
+
+declare global {
+    interface Window {
+        electronAPI: {
+            onUpdateLinks: (callback: (links: string) => void) => void;
+        };
+    }
+}
