+++ index.html
@@ -19,7 +19,7 @@
         </div>
     </div>
 
-    <script src="./renderer.js"></script>
+    <script src="./dist/renderer.js"></script>
 </body>
 
 </html>
