+++ package.json
@@ -11,12 +11,12 @@
     "url": "git+https://github.com/RewstApp/agent-smith-tray.git"
   },
   "license": "ISC",
-  "author": "",
+  "author": "Mikhael Glen Lataza",
   "type": "commonjs",
-  "main": "main.js",
+  "main": "dist/main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "electron ."
+    "start": "tsc && electron ."
   },
   "dependencies": {
     "@emotion/react": "^11.14.0",
@@ -25,6 +25,7 @@
     "@mui/material": "^7.3.2"
   },
   "devDependencies": {
-    "electron": "^38.1.0"
+    "electron": "^38.1.0",
+    "typescript": "^5.9.2"
   }
 }
