+++ src/preload.ts
@@ -1,6 +1,6 @@
-const { contextBridge, ipcRenderer } = require("electron/renderer");
+import { contextBridge, ipcRenderer } from "electron/renderer";
 
 contextBridge.exposeInMainWorld("electronAPI", {
-  onUpdateLinks: (callback) =>
+  onUpdateLinks: (callback: (value: string) => void) =>
     ipcRenderer.on("agent:links", (_event, value) => callback(value)),
 });
